+++ include/ILabel.h
@@ -33,7 +33,7 @@ class ILabel
     public:
         virtual ~ILabel() {}
 
-        virtual unsigned int id() const = 0;
+        virtual int64_t id() const = 0;
         virtual const std::string& name() const = 0;
         virtual std::vector<MediaPtr> files() = 0;
 };
